+++ test/ie.Tests.ps1
@@ -7,14 +7,16 @@ if (-not (Test-Path Variable:IsWindows)) { $script:IsWindows = $true }
 if (-not (Test-Path Variable:IsCoreCLR)) { $script:IsCoreCLR = $false }
 
 # Force-(re)import this module.
-Remove-Module -ea Ignore -Force (Split-Path -Leaf $PSScriptRoot)
-# Target the *.psd1 file explicitly, so the tests can run from versioned subfolders too. Note that the ModuleInfo's .Path property will reflect the *.psm1 instead.
-Import-Module (Get-Item $PSScriptRoot/../*.psd1)
+# Target the *.psd1 file explicitly, so the tests can run from versioned subfolders too. Note that the
+# loaded module's ModuleInfo's .Path property will reflect the *.psm1 instead.
+$manifest = (Get-Item $PSScriptRoot/../*.psd1)
+Remove-Module -ea Ignore -Force $manifest.BaseName # Note: To be safe, we unload any modules with the same name first (they could be in a different location and end up side by side in memory with this one.)
+Import-Module $manifest
 
 Describe 'ie tests' {
 
   It 'Only permits calls to external executables' {
-    { ie whoami } | Should -Not -Throw #  -ErrorId InvalidCommandType
+    { ie whoami } | Should -Not -Throw
     { ie Get-Date } | Should -Throw -ErrorId InvalidCommandType
     { ie select } | Should -Throw -ErrorId InvalidCommandType
     { ie help } | Should -Throw -ErrorId InvalidCommandType
@@ -30,11 +32,33 @@ Describe 'ie tests' {
     # !! in the behind-the-scenes command line, which ends up passing *3* arguments - there is no way to workarond that.
     # !! To make the test succeed, we use '3 " of snow' (space before ") in WinPS, in which case the engine does
     # !! double-quote the argument as a whole.
-    $exeArgs = '', 'a&b', ('3 " of snow', '3" of snow')[$IsCoreCLR], 'Nat "King" Cole', 'c:\temp 1\', 'a \" b'
+    $exeArgs = '', 'a&b', ('3 " of snow', '3" of snow')[$IsCoreCLR], 'Nat "King" Cole', 'c:\temp 1\', 'a \" b', 'a"b'
 
-    $result = dbea -Raw -UseIe $exeArgs
+    $result = dbea -Raw -UseIe -- $exeArgs
 
     Compare-Object $exeArgs $result | ForEach-Object { '{0} <{1}>' -f $_.SideIndicator, $_.InputObject } | Should -BeNull
+  }
+
+  It 'Properly passes scripts with complex quoting to various interpreters (if installed)' {
+    $ohtCmds = [ordered] @{
+      # CLIs that require \" and are escaped that way in both editions.
+      ruby       = { ie ruby -e 'puts "hi there"' }
+      perl       = { ie perl -E 'say "hi there"' }
+      pwsh       = { ie pwsh -noprofile -c '"hi there"' }
+      powershell = { ie powershell -noprofile -c '"hi there"' }
+
+      # CLIs that also accept "" and are used with that escaping in *WinPS*
+      node       = { ie node -pe '"hi there"' }
+      python     = { ie python -c 'print("hi there")' }
+    }
+
+    foreach ($exe in $ohtCmds.Keys) {
+      if (Get-Command -ea Ignore -Type Application $exe) {
+        "Testing with $exe...." | Write-Verbose -vb
+        & $ohtCmds[$exe] | Should -BeExactly 'hi there'
+      }
+    } 
+
   }
 
 }